+++ benchmark/Bench/Writer.hs
@@ -4,6 +4,7 @@ module Bench.Writer
 ( benchmark
 ) where
 
+import Control.Carrier.Writer.Church as C.Church
 import Control.Carrier.Writer.Strict as C.Strict
 import Control.Monad (replicateM_)
 #if MIN_VERSION_transformers(0,5,6)
@@ -18,7 +19,8 @@ benchmark :: Gauge.Benchmark
 benchmark = bgroup "Writer"
   [ bench "(,) w" $ whnf (fst . (tellLoop :: Int -> (Sum Int, ()))) n
   , bgroup "Identity"
-    [ bench "Strict.WriterC" $ whnf (run . C.Strict.execWriter @(Sum Int) . tellLoop) n
+    [ bench "Church.WriterC" $ whnf (run . C.Church.execWriter @(Sum Int) . tellLoop) n
+    , bench "Strict.WriterC" $ whnf (run . C.Strict.execWriter @(Sum Int) . tellLoop) n
 #if MIN_VERSION_transformers(0,5,6)
     , bench "CPS.WriterT"    $ whnf (run . T.CPS.execWriterT @_ @(Sum Int) . tellLoop) n
 #endif
@@ -26,7 +28,8 @@ benchmark = bgroup "Writer"
     , bench "Strict.WriterT" $ whnf (run . T.Strict.execWriterT @_ @(Sum Int) . tellLoop) n
     ]
   , bgroup "IO"
-    [ bench "Strict.WriterC" $ whnfAppIO (C.Strict.execWriter @(Sum Int) . tellLoop) n
+    [ bench "Church.WriterC" $ whnfAppIO (C.Church.execWriter @(Sum Int) . tellLoop) n
+    , bench "Strict.WriterC" $ whnfAppIO (C.Strict.execWriter @(Sum Int) . tellLoop) n
 #if MIN_VERSION_transformers(0,5,6)
     , bench "CPS.WriterT"    $ whnfAppIO (T.CPS.execWriterT @_ @(Sum Int) . tellLoop) n
 #endif